+++ MojFramework/Projectile.cpp
@@ -1,6 +1,6 @@
 #include "Projectile.h"
 
-Projectile::Projectile(const Vec2& pos_in, const Vec2& vel_in, float width_in, float height_in, float speed_in, float damage_in)
+Projectile::Projectile(const Vec2& pos_in, const Vec2& vel_in, float width_in, float height_in, float speed_in, int damage_in)
 	:
 	GameObject(pos_in, vel_in, width_in, height_in),
 	speed(speed_in),
@@ -46,7 +46,7 @@ bool Projectile::SmashedStatus()
 	return smashed;
 }
 
-float Projectile::DamageBonus()
+int Projectile::DamageBonus()
 {
 	return damage;
 }
